renamed modal to searchmodal
