+++ PlayFlix-firebase-FrontEnd-main/src/components/modal/SearchModal.js
